Added tests for the ACID and modified tests to work with the new blockId. Migration is now fully tested in the EE.

+++ tests/frontend/edu/brown/hstore/TestAntiCacheManager.java
@@ -360,7 +360,7 @@ public class TestAntiCacheManager extends BaseTestCase {
 
     @Test
     public void testReadNonExistentBlock() throws Exception {
-        short block_ids[] = new short[]{ 1111 };
+        int block_ids[] = new int[]{ 1111 };
         int tuple_offsets[] = new int[]{0}; 
         boolean failed = false;
         try {
